feat(admin): admin meetings page + copyable meeting link (#661)
Admin↔mentor parity: admins can now schedule/see meetings from their own
UI, and every meeting's link is one-click copyable (mentors benefit too).

- Extract the mentor meetings screen into a shared <MeetingsManager>; mount
  it at both /mentor/meetings and /admin/meetings. No role branching needed —
  /api/mentorship and /api/meetings already return the caller's own data for
  a MENTOR and everything for an ADMIN (meetings API already allowed ADMIN).
- Add a "Copy link" button + visible join link on each meeting row, with a
  clipboard-blocked prompt() fallback and a transient "Copied!" state.
- AdminNav: new Meetings entry (Video icon).
- i18n meetings.copyLink / linkCopied (EN/TR/DE).

First slice of the parity story #661; interaction-logging and targeted
email admin entry points remain tracked there.

Refs #661
