+++ src/i18n/dictionaries.ts
@@ -626,6 +626,8 @@ const en = {
     upcoming: 'Meetings',
     none: 'No meetings yet',
     scheduledCount: 'Invite sent to {n} mentee(s).',
+    copyLink: 'Copy link',
+    linkCopied: 'Copied!',
     pending: 'Pending',
     accepted: 'Accepted',
     declined: 'Declined',
@@ -2053,6 +2055,8 @@ const tr: Dict = {
     upcoming: 'Toplantılar',
     none: 'Henüz toplantı yok',
     scheduledCount: '{n} mentee’ye davet gönderildi.',
+    copyLink: 'Bağlantıyı kopyala',
+    linkCopied: 'Kopyalandı!',
     pending: 'Bekliyor',
     accepted: 'Kabul edildi',
     declined: 'Reddedildi',
@@ -3478,6 +3482,8 @@ const de: Dict = {
     upcoming: 'Meetings',
     none: 'Noch keine Meetings',
     scheduledCount: 'Einladung an {n} Mentee(s) gesendet.',
+    copyLink: 'Link kopieren',
+    linkCopied: 'Kopiert!',
     pending: 'Ausstehend',
     accepted: 'Zugesagt',
     declined: 'Abgesagt',
